backend/requirements.txt
fastapi
uvicorn

# for databases
sqlalchemy
psycopg2

# for env var
python-dotenv

# for email validation
pydantic[email]

# hashing
passlib[bcrypt]

# test
requests
pytest
pytest-cov

# for JWT token
python-jose